requirements.txt
# Google Drive
google-auth>=2.28.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# OpenRouter / LLM
openai>=1.30.0

# Document generation
python-docx>=1.1.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0

# PDF / DOC parsing
pdfplumber>=0.11.0

# Utilities
python-dotenv>=1.0.0
